Move BotCard and BotCategory components to page-specific components folder
