+++ frontend/src/pages/protected/SelectBot/SelectBot.tsx
@@ -1,6 +1,6 @@
-import { bots as botsTable } from "../../constants/botsConfig";
-import DashboardNavbar from "../../components/page/dashboard/DashboardNavbar";
-import BotCategory from "../../components/page/playBot/BotCategory";
+import { bots as botsTable } from "../../../constants/botsConfig";
+import DashboardNavbar from "../../../components/page/dashboard/DashboardNavbar";
+import BotCategory from "./components/BotCategory";
 
 import "../../styles/pages/select-bot.scss";
 
@@ -11,11 +11,14 @@ function SelectBot() {
 			<div className="main-page-content">
 				<h1 className="select-bot-page-heading">Play vs bot</h1>
 				<p className="select-bot-page-description">
-                    Play against a bot of your choice
-                </p>
-                <div className="bots-container">
+					Play against a bot of your choice
+				</p>
+				<div className="bots-container">
 					{botsTable.map((categoryInfo, _) => (
-						<BotCategory key={categoryInfo.categoryId} categoryInfo={categoryInfo} />
+						<BotCategory
+							key={categoryInfo.categoryId}
+							categoryInfo={categoryInfo}
+						/>
 					))}
 				</div>
 			</div>
