+++ frontend/src/pages/protected/SelectBot/components/BotCard.tsx
@@ -1,8 +1,8 @@
 import { useNavigate } from "react-router-dom";
-import { BotInfo } from "../../../constants/botsConfig";
+import { BotInfo } from "../../../../constants/botsConfig";
 
 import "../../../styles/features/playBot/bot-card.scss";
-import { createBotGame } from "../../../utils/apiUtils";
+import { createBotGame } from "../../../../utils/apiUtils";
 
 type BotCardProps = {
 	botInfo: BotInfo;
@@ -18,9 +18,9 @@ function BotCard({ botInfo }: BotCardProps) {
 			state: {
 				gameId: botGameId,
 				bot: botInfo.botId,
-				assignedColor: assignedColor
-			}
-		})
+				assignedColor: assignedColor,
+			},
+		});
 	}
 
 	function getBotImage() {
